+++ TroveQL/src/arc/arc.ts
@@ -1,32 +1,19 @@
 import { CacheItem } from './CacheItem';
+import {
+  RequestBody,
+  ItemType,
+  CacheType,
+  getResponse,
+  fetchResponse,
+} from './arcTypes';
 
-type ItemType = {
-  value: string;
-  hits: number;
-};
-//NEW SHAPE OF QUERY from Erika
-
-// type Variables = {
-//   [key: string]: string;
-// };
-// type RequestBody = {
-//   query: string;
-//   variables: Variables;
-// };
-
-type CacheClass = Map<string, ItemType>;
-type GhostClass = Map<string, boolean>;
-
-class TroveCache {
-  t1: CacheClass;
-  t2: CacheClass;
-  b1: GhostClass;
-  b2: GhostClass;
+export class TroveCache {
   capacity: number;
   p: number;
-  t1Size: number;
-  t2Size: number;
-  ghostSize: number;
+  t1: CacheType;
+  t2: CacheType;
+  b1: CacheType;
+  b2: CacheType;
 
   constructor(size: number) {
     this.capacity = size;
@@ -35,112 +22,130 @@ class TroveCache {
     this.t2 = new Map();
     this.b1 = new Map();
     this.b2 = new Map();
-    //we need to remove this
-    this.t1Size = Math.floor(this.capacity * this.p);
-    this.t2Size = this.capacity - this.t1Size;
-    this.ghostSize = this.t1Size;
   }
 
-  // RETRIEVE ITEMS FROM THE CACHE
-  public get = (query: string): string | boolean => {
-    const cache = this.getCache(query);
-    if (cache === undefined) {
-      this.resizeCache('miss');
-      return undefined;
-    }
-    if (cache === this.t1 || cache === this.t2) {
-      const result: ItemType = cache.get(query);
+  public get = (query: string): getResponse => {
+    switch (true) {
+      case this.t1.has(query) || this.t2.has(query):
+        console.log('In Get Case I');
+        const cache = this.t1.has(query) ? this.t1 : this.t2;
+        const result = cache.get(query) as ItemType; //this may be an issue
+        cache.delete(query);
+        this.t2.set(query, result);
 
-      //get item from T1 & promote it to T2
-      const isT1 = cache === this.t1;
-      if (isT1) {
-        this.t1.delete(query);
-      } else {
-        this.t2.delete(query);
-      }
-      //TODO: change to LFU evict when logic is build
-      if (this.t2.size >= this.t2Size) {
-        this.evictLRU(this.t2);
-      }
+        return {
+          result: result.value,
+          miss: false,
+        };
+        break;
 
-      result.hits++;
-      this.t2.set(query, result);
-      return result.value;
-    } else if (cache === this.b1 || cache === this.b2) {
-      this.resizeCache(cache);
-      return false;
+      case this.b1.has(query):
+        console.log('In Get Case II');
+        return { result: '', miss: 'b1' };
+        break;
+
+      case this.b2.has(query):
+        console.log('In Get Case III');
+        return { result: '', miss: 'b2' };
+        break;
+
+      case !this.t1.has(query) &&
+        this.t2.has(query) &&
+        this.b1.has(query) &&
+        this.b2.has(query):
+        console.log('In Get Case III');
+        return { result: '', miss: 'miss' };
+        break;
     }
   };
 
-  //ADD NEW ITEMS TO THE CACHE
-  public set = (query: string, result: string): void => {
-    const newItem = new CacheItem(result);
-    if (this.t1.size >= this.t1Size) {
-      this.evictLRU(this.t1);
+  public set = (res: fetchResponse) => {
+    const node = new CacheItem(res.result);
+
+    switch (true) {
+      case res.miss === 'b1':
+        console.log('In Set Case II');
+        this.adaptation(true);
+        this.replace(false);
+
+        this.t2.set(res.query, node);
+        break;
+
+      case res.miss === 'b2':
+        console.log('In Set Case III');
+        this.adaptation(false);
+        this.replace(true);
+
+        this.t2.set(res.query, node);
+        break;
+
+      case res.miss === 'miss':
+        console.log('In Set Case III');
+        const l1: number = this.t1.size + this.t2.size;
+
+        switch (true) {
+          case l1 === this.capacity:
+            if (this.t1.size < this.capacity) {
+              this.evictLRU(this.b1);
+              this.replace(false); //check this
+            } else {
+              this.evictLRU(this.t1);
+            }
+            break;
+
+          case l1 < this.capacity:
+            const totalSize =
+              this.t1.size + this.b1.size + this.t2.size + this.b2.size;
+            if (totalSize >= this.capacity) {
+              if (totalSize === this.capacity * 2) {
+                this.evictLRU(this.b2);
+              }
+              this.replace(false); //check this
+            }
+            break;
+        }
+        //add data
+        this.t1.set(res.query, node);
+
+        break;
     }
-    this.t1.set(query, newItem);
   };
 
-  //DELETE ALL ITEMS FROM ALL GHOSTS AND CACHES
-  public deleteAll = (): void => {
-    const caches = [this.t1, this.t2, this.b1, this.b2];
+  private adaptation = (isB1: boolean) => {
+    if (isB1) {
+      const n: number =
+        this.b1.size >= this.b2.size ? 1 : this.b2.size / this.b1.size;
+      this.p = this.p + n;
+    } else {
+      const n: number =
+        this.b2.size >= this.b1.size ? 1 : this.b1.size / this.b2.size;
+      this.p = this.p - n;
+    }
+  };
+
+  private evictLRU(cache: CacheType) {
+    const firstKey = cache.keys().next().value;
+    // will the following line pass by ref? Do we need to make a shallow copy?
+    const evicted = [firstKey, cache.get(firstKey)];
+    cache.delete(firstKey);
+    return evicted;
+  }
+
+  private replace(foundInB2: boolean): void {
+    if (
+      this.t1.size > 0 &&
+      (this.t1.size > this.p || (foundInB2 && this.t1.size === this.p))
+    ) {
+      let [key, cacheItem] = this.evictLRU(this.t1);
+      this.b1.set(key, cacheItem);
+    } else {
+      let [key, cacheItem] = this.evictLRU(this.t2);
+      this.b2.set(key, cacheItem);
+    }
+  }
+
+  public removeAll = (): void => {
+    const caches: CacheType[] = [this.t1, this.t2, this.b1, this.b2];
     caches.forEach((cache) => cache.clear());
   };
-
-  //HELPER METHODS
-  private getCache(query: string): CacheClass | GhostClass | undefined {
-    const caches = [this.t1, this.t2, this.b1, this.b2];
-    for (let cache of caches) {
-      if (cache.has(query)) return cache;
-    }
-    return undefined;
-  }
-
-  private evictLRU(cache: CacheClass | GhostClass): void {
-    const firstKey = cache.keys().next().value;
-    let targetCache;
-    if (cache === this.t1 || cache === this.t2) {
-      targetCache = cache === this.t1 ? this.b1 : this.b2;
-    }
-    if (targetCache && targetCache.size >= this.ghostSize) {
-      this.evictLRU(targetCache);
-    }
-    cache.delete(firstKey);
-  }
-
-  //TODO: Write logic for LFU cache eviction
-  private evictLFU(cache: CacheClass | GhostClass): void {}
-
-  private resizeCache(cache: GhostClass | 'miss'): void {
-    const growCacheSize =
-      cache === this.b1 || cache === 'miss' ? this.t1Size : this.t2Size;
-    const targetCache =
-      cache === this.b1 || cache === 'miss' ? this.t1 : this.t2;
-    const shrinkCacheSize = targetCache === this.t1 ? this.t2Size : this.t1Size;
-
-    this.updatePValue(growCacheSize, shrinkCacheSize);
-
-    while (targetCache.size > growCacheSize) {
-      this.evictLRU(targetCache);
-    }
-  }
-  //update pValue Helper function for resizeCache()
-  private updatePValue(growCacheSize: number, shrinkCacheSize: number): void {
-    //rethink this to increment pValue by a percentage
-    if (shrinkCacheSize <= 1) {
-      throw new Error('Cache sizes cannot be zero or negative');
-      return;
-    }
-    //this needs to be refactored
-    shrinkCacheSize--;
-    growCacheSize++;
-    const [min, max] = [
-      Math.min(growCacheSize, shrinkCacheSize),
-      Math.max(growCacheSize, shrinkCacheSize),
-    ];
-    console.log(max, min);
-    this.p = max / min;
-  }
 }
-
-export { TroveCache };
